before merge with Ye

+++ Code/Android/GameTrade-in/app/src/main/java/com/example/ye/gametrade_in/fragment/PaginationFragment.java
@@ -75,7 +75,7 @@ public abstract class PaginationFragment<T> extends Fragment implements Paginati
 
     GameTradeService mGameTradeService;
 
-    protected int mUserId;
+    protected Long mUserId;
 
     protected abstract Call<List<T>> callApi() ;
 
@@ -90,7 +90,7 @@ public abstract class PaginationFragment<T> extends Fragment implements Paginati
                     QueryPreferences.getStoredAuthorizedQuery(this.getActivity().getApplicationContext());
 
         if(QueryPreferences.getStoredUserIdQuery(this.getActivity().getApplicationContext()) != null){
-            mUserId = Integer.parseInt(
+            mUserId = Long.parseLong(
                         QueryPreferences.getStoredUserIdQuery(this.getActivity().getApplicationContext())
                 );
         }
