+++ Code/Android/GameTrade-in/app/src/main/java/com/example/ye/gametrade_in/fragment/ReloadableFragment.java
@@ -64,7 +64,7 @@ public abstract class ReloadableFragment<T> extends Fragment implements Paginati
 
     GameTradeService mGameTradeService;
 
-    protected int mUserId;
+    protected Long mUserId;
 
     protected abstract Call<List<T>> callApi() ;
 
@@ -79,7 +79,7 @@ public abstract class ReloadableFragment<T> extends Fragment implements Paginati
                 QueryPreferences.getStoredAuthorizedQuery(this.getActivity().getApplicationContext());
 
         if(QueryPreferences.getStoredUserIdQuery(this.getActivity().getApplicationContext()) != null){
-            mUserId = Integer.parseInt(
+            mUserId = Long.parseLong(
                     QueryPreferences.getStoredUserIdQuery(this.getActivity().getApplicationContext())
             );
         }
